+++ malachite-q/src/test_util/generators/special_random.rs
@@ -93,6 +93,33 @@ pub fn special_random_gaussian_rational_pair_gen(
     )))
 }
 
+pub fn special_random_gaussian_rational_pair_gen_var_1(
+    config: &GenConfig,
+) -> It<(GaussianRational, GaussianRational)> {
+    Box::new(random_pairs(
+        EXAMPLE_SEED,
+        &|seed| {
+            striped_random_gaussian_rationals(
+                seed,
+                config.get_or("mean_stripe_n", 32),
+                config.get_or("mean_stripe_d", 1),
+                config.get_or("mean_bits_n", 64),
+                config.get_or("mean_bits_d", 1),
+            )
+        },
+        &|seed| {
+            striped_random_gaussian_rationals(
+                seed,
+                config.get_or("mean_stripe_n", 32),
+                config.get_or("mean_stripe_d", 1),
+                config.get_or("mean_bits_n", 64),
+                config.get_or("mean_bits_d", 1),
+            )
+            .filter(|x| x.real != 0u32 || x.imaginary != 0u32)
+        },
+    ))
+}
+
 pub fn special_random_gaussian_rational_triple_gen(
     config: &GenConfig,
 ) -> It<(GaussianRational, GaussianRational, GaussianRational)> {