+++ malachite-q/tests/lib.rs
@@ -70,6 +70,7 @@ pub mod gaussian_rational {
         pub mod abs_squared;
         pub mod add;
         pub mod conjugate;
+        pub mod div;
         pub mod div_i;
         pub mod is_power_of_2;
         pub mod mul;